refactor: PriceInputコンポーネントをuseLensを使用してリファクタリングし、関連するハンドラーを追加
