--- app/components/forms/PriceInput/types.ts
@@ -1,9 +1,0 @@
-/**
- * NOTE:
- * 循環参照になるためtypes.tsに分離している
- */
-export type CustomProps = {
-  value: number
-  onChange: (value?: number) => void
-  onBlur: () => void
-}
